requirements.txt
acme==2.9.0
annotated-types==0.7.0
anyio==4.10.0
certbot==2.9.0
certbot-nginx==2.9.0
certifi==2023.11.17
chardet==5.2.0
ConfigArgParse==1.7
configobj==5.0.8
cryptography==41.0.7
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.6
jiter==0.10.0
josepy==1.14.0
openai==1.106.1
parsedatetime==2.6
pydantic==2.11.7
pydantic_core==2.33.2
PyICU==2.12
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyRFC3339==1.1
python-dotenv==1.1.1
pytz==2024.1
requests==2.31.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.0.7
